small improvement & fixes.
- fix the max number of landmark selected.
- rework the selection of landmark in ALT preprocessing to start by the smallest node ID.
- fix getOneOrTwo issue who take all integrer number in entry.
- fix API only to take poitive integrer and return error if not the exact type. (for example : 1ABC will prune to 1, but now, it will just return an error "INVALID PARAMETER").
- return an error when the graph is empty at preprocessing time.

+++ src/decl.h
@@ -177,7 +177,7 @@ void clear();
 int getInteger(const std::string& prompt);
 double getPercentage(const std::string& prompt);
 bool getYesNo(const std::string& prompt);
-double getOneOrTwo(const std::string& prompt);
+int getOneOrTwo(const std::string& prompt);
 
 void initializeSearchBuffers(const graph& gdata, search_buffers& buffers);
